NN model layouts can now be loaded from disk
In order to avoid hard-coding of the network layout, the model is now a positional parameter.
Since the models stored in the JSON files specify their input_shape, the resolution is now determined by them, so the -r --resolution parameter has been removed.
One can also execute the nn.py file directy to write some predefined models to the disk.